+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Do not exit when a YAML config has no sonarr or radarr section.
 - Sonarr: Invalid release profile JSON files no longer cause the program to exit. Instead, it just
   skips them and prints a warning to the user. (#87)
+- Radarr: Reword the warning about missing scores for CFs to make it more clear that having no score
+  does not prevent CFs from being synced.
 
 ## [2.2.0] - 2022-06-03